Refatora persistência de dispositivos para camada de serviço
Toda a lógica de persistência de dispositivos foi extraída do DeviceController e movida para a nova camada DeviceService, exposta via IDeviceService. O controller agora injeta e utiliza o serviço, tornando o código mais modular e testável. Foram criados arquivos dedicados para respostas de upsert em lote (UpsertDeviceResponse) e adicionados os novos arquivos ao projeto. Essa refatoração melhora a separação de responsabilidades e facilita futuras manutenções e extensões.
