+++ Source/Libraries/openPDC.Adapters/openPDC.Adapters.csproj
@@ -115,6 +115,8 @@
     <Compile Include="GrafanaController.cs" />
     <Compile Include="PhasorController.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
+    <Compile Include="Services\DeviceService.cs" />
+    <Compile Include="Services\IDeviceService.cs" />
   </ItemGroup>
   <ItemGroup>
     <ProjectReference Include="..\openPDC.Model\openPDC.Model.csproj">
