+++ Source/Libraries/openPDC.Model/openPDC.Model.csproj
@@ -68,6 +68,7 @@
     <Compile Include="Protocol.cs" />
     <Compile Include="SchemaVersion.cs" />
     <Compile Include="StatusLight.cs" />
+    <Compile Include="UpsertDeviceResponse.cs" />
   </ItemGroup>
   <Import Project="$(MSBuildToolsPath)\Microsoft.CSharp.targets" />
   <!-- To modify your build process, add your task inside one of the targets below and uncomment it.
